Move stdin_file_input to a separate header file
