+++ libvast/test/loader_plugin.cpp
@@ -11,15 +11,16 @@
 #include "vast/pipeline.hpp"
 #include "vast/plugin.hpp"
 #include "vast/table_slice.hpp"
+#include "vast/test/stdin_file_inut.hpp"
 #include "vast/test/test.hpp"
 
+#include <arrow/record_batch.h>
 #include <caf/error.hpp>
 #include <caf/test/dsl.hpp>
 
-#include <fcntl.h>
-#include <unistd.h>
-
 using namespace vast;
+
+using test::stdin_file_input;
 
 namespace {
 
@@ -70,28 +71,6 @@ struct fixture {
   const vast::loader_plugin* loader_plugin;
   std::function<auto()->generator<chunk_ptr>> current_loader;
   mock_control_plane control_plane;
-};
-
-// Helper struct that, as long as it is alive, redirects stdin to the output of
-// a file.
-template <detail::string_literal FileName = "">
-struct stdin_file_input {
-  stdin_file_input() {
-    old_stdin_fd = ::dup(fileno(stdin));
-    REQUIRE_NOT_EQUAL(old_stdin_fd, -1);
-    auto input_file_fd = ::open(
-      fmt::format("{}{}", VAST_TEST_PATH, FileName.str()).c_str(), O_RDONLY);
-    REQUIRE_NOT_EQUAL(input_file_fd, -1);
-    ::dup2(input_file_fd, fileno(stdin));
-    ::close(input_file_fd);
-  }
-
-  ~stdin_file_input() {
-    auto old_stdin_status = ::dup2(old_stdin_fd, fileno(stdin));
-    REQUIRE_NOT_EQUAL(old_stdin_status, -1);
-    ::close(old_stdin_fd);
-  }
-  int old_stdin_fd;
 };
 
 } // namespace
